fix(daemon): use uint64 for iptables counters to prevent overflow
iptables packet/byte counters are kernel uint64 values but were stored
as int and parsed with strconv.Atoi. When counters exceed MaxInt64 the
parse fails silently. Change GwIPTablesCounters fields to uint64 and
use strconv.ParseUint for correct handling of large counter values.

+++ pkg/util/iptables.go
@@ -9,6 +9,6 @@ type IPTableRule struct {
 }
 
 type GwIPTablesCounters struct {
-	Packets     int
-	PacketBytes int
+	Packets     uint64
+	PacketBytes uint64
 }
